Broken Autos
Path following doesn't work, no matter whether you run a path in an auto or separately. In networktables I can see that within the pathplanner tab our target pose is changing, but our actual pose doesnt change to reflect that. I tried using the set chassisSpeeds method I gave the autobuilder and it works when bound to a button, so idk what is wrong (yet)